Refactor package of scheduling classess
