+++ app/src/main/java/com/gmail/brunokawka/poland/sleepcyclealarm/ui/tabs/accessalarm/alarms/AlarmsAdapter.java
@@ -9,7 +9,6 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.gmail.brunokawka.poland.sleepcyclealarm.R;
-import com.gmail.brunokawka.poland.sleepcyclealarm.alarm.AlarmController;
 import com.gmail.brunokawka.poland.sleepcyclealarm.data.pojo.Alarm;
 
 import butterknife.BindView;
